Change the code that avoids recursively backing up the snapshot root to use an anchored absolute-path exclude rule.
Suppose the backup point is /foo and the snapshot root is /foo/sr .
When --relative is on, the existing code would use --exclude=sr ,
which would incorrectly match files such as /foo/bar/sr .  See:

https://qa.mandriva.com/show_bug.cgi?id=45979

We now use an anchored exclude rule with the absolute-path modifier:
--filter='-/ /foo/sr' .  This avoids matching unrelated files, and as a
bonus, it works regardless of --relative, allowing us to retire the code
path that ran rsync once for each child of the backup point.  It should
work correctly in any rsync version since 2.6.4, except for 2.6.7.