Revert "Detect busybox logger and change logger cmdline"
This reverts commit 0a056977d736df3adcc16e99a50dd06efbdb7a0e.

That commit is reverted because from now on we will print the
rnapshot's current pid directly in the tag instead of using the
"-i" option of the logger program. Using the "-i" option prints out
the PID of the logger program itself (therefore making it impossible to
correlate log lines and rsnapshot instances when there are multiple ones
running simultaneously).