Print rnspashot's PID when logging to syslog
Print rsnapshot's PID instead of the logger's PID when logging to
syslog. This is much more useful because it allow to relate log
lines to rsnapshot instances when there are multiple concurrent
ones. Previously each line had a different PID, even those
generated by the same instance.

The change is achieved by dropping the "-i" option from the
logger's call and appending the rnapshot's PID to the tag itself.