Fix error when file matching "/rsync*" in backup.
The script synchronises with the output from rsnapshot by looking for the command line invoking rsync. The regex is too loose, so a false match will occur if any file matching "/rsync*" is in the backup - this means that backup of the root filesystem will always error when it gets to the rsync binary. The change tightens the regex to reduce the chance of a false match.