strip redundant trailing slashes from backup dest
After 2a2b302 (Fix regression in 262 (#283), 2021-06-08), the newly
added normalize_dest_file_path_part() function changes a backup
destination of './' or '.' to '/'.

This results in rsync_backup_point() adding a redundant trailing '/' to
the rsync destination (and --link-dest argument, if enabled).

The redundant trailing slash isn't harmful, but it looks odd in the
output and logs.  Remove it while generating the rsync command stack.